+++ apps/admin_web/lib/app/admin_shell.dart
@@ -2,6 +2,7 @@ import 'package:admin_web/features/content/presentation/content_hub_page.dart';
 import 'package:admin_web/features/moderation/presentation/asset_review_page.dart';
 import 'package:admin_web/features/platform/presentation/platform_dashboard_page.dart';
 import 'package:admin_web/features/schools/presentation/schools_page.dart';
+import 'package:admin_web/features/users/presentation/users_page.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:nano_design_system/nano_design_system.dart';
@@ -23,6 +24,7 @@ class AdminShell extends StatelessWidget {
     this.assetReviewRepository,
     this.platformDashboardRepository,
     this.schoolAdminRepository,
+    this.platformUserRepository,
   });
 
   final EnvironmentConfig config;
@@ -37,6 +39,7 @@ class AdminShell extends StatelessWidget {
   final AssetReviewRepository? assetReviewRepository;
   final PlatformDashboardRepository? platformDashboardRepository;
   final SchoolAdminRepository? schoolAdminRepository;
+  final PlatformUserRepository? platformUserRepository;
 
   @override
   Widget build(BuildContext context) {
@@ -148,6 +151,7 @@ class AdminDestinationPage extends StatelessWidget {
     this.assetReviewRepository,
     this.platformDashboardRepository,
     this.schoolAdminRepository,
+    this.platformUserRepository,
   });
 
   final NavDestination destination;
@@ -157,6 +161,7 @@ class AdminDestinationPage extends StatelessWidget {
   final AssetReviewRepository? assetReviewRepository;
   final PlatformDashboardRepository? platformDashboardRepository;
   final SchoolAdminRepository? schoolAdminRepository;
+  final PlatformUserRepository? platformUserRepository;
 
   @override
   Widget build(BuildContext context) {
@@ -174,6 +179,12 @@ class AdminDestinationPage extends StatelessWidget {
         principal.role == AppRole.superadmin &&
         schoolAdminRepository != null) {
       return SchoolsPage(repository: schoolAdminRepository!);
+    }
+
+    if (destination.id == 'users' &&
+        principal.role == AppRole.superadmin &&
+        platformUserRepository != null) {
+      return UsersPage(repository: platformUserRepository!);
     }
 
     if (destination.id == 'content' &&